+++ tests/102_projectstarter/004_init.py
@@ -48,7 +48,7 @@ def test_success(projectstarter, temp_builds_dir, name, ext, module,
             'TARGET_PATH': 'css',
             "LIBRARY_PATHS": [],
             "OUTPUT_STYLES": "nested",
-            "HASH_SUFFIXES": False,
+            "HASH_SUFFIX": None,
             "SOURCE_COMMENTS": False,
             "EXCLUDES": []
         }